moved config up
